+++ automation/.classpath
@@ -5,5 +5,6 @@
 	<classpathentry kind="src" path="src"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/javacv"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/systemhook"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/jnativehook"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
