+++ automation/.gitignore
