+++ automation/.settings/org.eclipse.jdt.core.prefs
