+++ automation/build.gradle
@@ -1,0 +1,73 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This generated file contains a sample Java project to get you started.
+ * For more details take a look at the Java Quickstart chapter in the Gradle
+ * User Manual available at https://docs.gradle.org/5.3/userguide/tutorial_java_projects.html
+ */
+buildscript {
+    dependencies {
+        classpath group: 'de.dynamicfiles.projects.gradle.plugins', name: 'javafx-gradle-plugin', version: '8.8.2'
+    }
+
+    repositories {
+        mavenLocal()
+        mavenCentral()
+    }
+}
+plugins {
+    // Apply the java plugin to add support for Java
+    id 'java'
+    // Apply the application plugin to add support for building an application
+    id 'application'
+}
+
+//javafx {
+//	version = "8.0.0"
+//	modules = [ 'javafx.controls', 'javafx.fxml']
+//}
+version = '1.0'
+mainClassName = 'automation.FXGUI'
+//makes a jar with dependencies
+jar {
+    manifest {
+        attributes "Main-Class": mainClassName
+    }
+
+    from {
+        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
+sourceSets.main {
+    java {
+        srcDir 'src/main/java' //assume that your source codes are inside this path
+    }
+    resources {
+        srcDirs = ['src/main/java', 'src/main/resources']
+        exclude "**/*.java"
+    }
+}
+apply plugin: 'javafx-gradle-plugin'
+jfx {
+	mainClass = mainClassName
+	vendor = 'AutoRoutine'
+}
+repositories {
+    // Use jcenter for resolving your dependencies.
+    // You can declare any Maven/Ivy/file repository here.
+    jcenter()
+	maven {
+		url "https://raw.github.com/kristian/system-hook/mvn-repo/"
+	}
+	mavenLocal()
+	mavenCentral()
+}
+
+dependencies {
+    // This dependency is found on compile classpath of this component and consumers.
+    implementation 'com.google.guava:guava:27.0.1-jre'
+	implementation 'com.1stleg:jnativehook:2.1.0'
+	implementation 'de.dynamicfiles.projects.gradle.plugins:javafx-gradle-plugin:8.8.2'
+    // Use JUnit test framework
+    testImplementation 'junit:junit:4.12'
+}
