+++ automation/settings.gradle
