+++ automation/src/main/java/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Student</name>
+	<name>java</name>
 	<comment></comment>
 	<projects>
 	</projects>
