+++ automation/src/main/java/automation/.gitignore
@@ -1,0 +1,6 @@
+/Auth.class
+/Auto.class
+/FXGUI.class
+/LoginController.class
+/MainController.class
+/NewUserController.class
