+++ automation/src/main/java/automation/Auth.java
@@ -1,4 +1,4 @@
-package automation;
+package main.java.automation;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
