+++ automation/src/main/java/automation/FxGUInewuser.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="automation.NewUserController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.java.automation.NewUserController">
    <children>
       <Button fx:id="createNewUser" layoutX="248.0" layoutY="344.0" mnemonicParsing="false" onAction="#createNewUser" text="Create New User" />
       <Label layoutX="232.0" layoutY="71.0" text="Create New User Account" />
