+++ automation/src/test/java/src/AppTest.java
@@ -1,13 +1,11 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package automation;
+package test.java.src;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
 
 public class AppTest {
-    @Test public void testAppHasAGreeting() {
-        //TODO make tests
-    }
+    //TODO make unit tests
 }
