--- src/main/java/automation/bin/.gitignore
@@ -1,1 +1,0 @@
-/automation/
